Move EnvelopeGenerator to synth/envelope.rs module
Refactor the EnvelopeGenerator and EnvelopeState out of synth.rs into a dedicated synth/envelope.rs submodule. Update all references and visibility as needed. This improves code organization and follows project code style guidelines by avoiding mod.rs and keeping modules flat unless grouping related submodules. No functional changes.
